more fix

+++ bot/backend/promo.py
@@ -10,14 +10,14 @@ async def create_fortune_wheel(request: Request):
     form_data = await request.json()
     print(form_data)
     ticket_name = form_data['ticket_name']
-    bonus = int(form_data['bonus'])
+    bonus = float(form_data['bonus'])
     type_promo = int(form_data['type'])
     number_of_uses = int(form_data['number_of_uses'])
     number_of_users = int(form_data['number_of_users'])
-    time_of_existence = int(form_data['time_of_existence']) * 3600 * 24
-    time_of_duration = int(form_data['time_of_duration']) * 3600 * 24
-    min_wager = int(form_data['min_wager'])
-    wager = int(form_data['wager'])
+    time_of_existence = float(form_data['time_of_existence']) * 3600 * 24
+    time_of_duration = float(form_data['time_of_duration']) * 3600 * 24
+    min_wager = float(form_data['min_wager'])
+    wager = float(form_data['wager'])
     special_users = form_data['special_users']
 
     type_promo = "balance" if type_promo == 1 else 'ticket'
